make meme bot a cog
