+++ main2.py
@@ -1,16 +1,19 @@
 # main.py
-from discord.ext import commands
+import asyncio
 import os
+
+import discord
+from discord.ext import commands
 from dotenv import load_dotenv
 
 load_dotenv()
 TOKEN = os.getenv('TOKEN')
 
-client = commands.Bot(command_prefix = "!")
+client = commands.Bot(command_prefix="!", intents=discord.Intents.all())
 
 # Looks inside the /cogs/ folder and loads up all of our cogs
 for filename in os.listdir("./cogs"):
-	if filename.endswith(".py"):
-		client.load_extension("cogs." + filename[:-3])
+    if filename.endswith(".py"):
+        asyncio.run(client.load_extension(f"cogs.{filename[:-3]}"))
 
 client.run(TOKEN)
